added --check

+++ sxmo/homestatus_widget.sh
@@ -1,5 +1,10 @@
 #!/usr/bin/env sh
-. /etc/os-release
+if [ "$1" = "--check" ]; then
+    pidof wayout || exit 0
+    export SWAYSOCK="$(cat "$CACHEDIR"/sxmo/sxmo.swaysock)"
+    export XDG_RUNTIME_DIR=~/.local/run/
+    export CACHEDIR="~/.cache/"
+fi
 FONTSIZE="22"
 pgrep -f notifyclient.sh || ~/dotfiles/notifyclient.sh &
 pgrep -f homestatus.sh && pkill -f homestatus.sh
